+++ NEWS.md
@@ -27,6 +27,8 @@ shiny 1.5.0.9000
 
 ### Minor new features and improvements
 
+* Make plot resolution dynamic, so it can be changed in server like width and height.
+
 * When UI is specified as a function (e.g. `ui <- function(req) { ... }`), the response can now be an HTTP response as returned from the (newly exported) `httpResponse()` function. (#2970)
 
 * `selectInput` and `selectizeInput` now warn about performance implications when thousands of choices are used, and recommend [server-side selectize](https://shiny.rstudio.com/articles/selectize.html) be used instead. (#2959)